Add :results file for org mode R code template
NB: as of org version 9.3, the :results argument must include both file and graphics; earlier versions only required the graphics argument
https://orgmode.org/worg/org-contrib/babel/languages/ob-doc-R.html#org04ba7ed
https://orgmode.org/Changes.html

The "file" :results value is now mandatory for a code block
returning a link to a file.  The :file or :file-ext header
arguments no longer imply a "file" result is expected.
https://code.orgmode.org/bzg/org-mode/commit/26ed66b23335eb389f1f2859e409f46f66279e15